fix redux tests
